Included the secret in the manual and automated workflows (#20)
* Included the secret in the manual and automated workflows; fixed the envs not appearing in the pods bug

* Typo in workflow

+++ deploy_no_unity_build.sh
@@ -4,6 +4,10 @@
 # Must place the Unity build in build/WebGL/WebGL e.g. build/WebGL/WebGL/index.html must exist
 ######################
 
+######################
+# Must define the env FIRESTORE_KEY or have a file called FIRESTORE_KEY with the Firestore authentication key
+######################
+
 # Needs docker and kubectl installed and set up
 # Docker: https://docs.docker.com/get-docker/
 # Google SDK: https://cloud.google.com/sdk/docs/install#linux
@@ -18,6 +22,19 @@ PROJECT_ID=temptool
 IMAGE=temptool
 DEPLOYMENT_NAME=temptool
 TAG=$(uuidgen)
+
+# Check if FIRESTORE_KEY is defined and create a file with that name
+echo "Checking if FIRESTORE_KEY is defined as env"
+if [[ -v FIRESTORE_KEY ]]
+then
+    echo "FIRESTORE_KEY env exists, writing its contents to file FIRESTORE_KEY..."
+    rm -rf FIRESTORE_KEY
+    printenv FIRESTORE_KEY > FIRESTORE_KEY || echo "Failed to copy the contents of the FIRESTORE_KEY env variable into a file with the same name"
+else
+    echo "FIRESTORE_KEY env does not exist; file FIRESTORE_KEY must exist to not fail"
+fi
+
+test -f FIRESTORE_KEY || { echo "File FIRESTORE_KEY does not exist"; exit 1; }
 
 # Save kustomization.yaml
 cp kustomization.yaml __kustomization.yaml || { echo "Failed to backup old kustomization.yaml file"; exit 1; }
@@ -35,6 +52,7 @@ docker push "gcr.io/$PROJECT_ID/$IMAGE:$TAG" || { echo "Failed to push docker im
 # Deploy
 echo "#### Deploying to kubernetes ####"
 ./kustomize edit set image gcr.io/temptool/temptool=gcr.io/$PROJECT_ID/$IMAGE:$TAG || { echo "Failed to modify the kustomize file to use the newly built docker image"; exit 1; }
+./kustomize edit add secret firestore-key --from-file=FIRESTORE_KEY || { echo "Failed to modify the kustomize file to add the firestore auth"; exit 1; }
 ./kustomize build . | kubectl apply -f - || { echo "Failed to deploy"; exit 1; }
 kubectl rollout status deployment/$DEPLOYMENT_NAME || echo "Failed to watch deployment"
 kubectl get services -o wide || echo "Failed to get kubernetes services"
